pipeline/requirements.txt
anthropic>=0.40.0
google-generativeai>=0.8.0
elevenlabs>=1.0.0
openai>=1.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
requests>=2.31.0
